Update README.md
improved and added changes to the code. now the user can change the schedule of any specific day, compared to the previous code in which you had to fill all the slots of the previous day in order to proceed to the next day. Also added destructors to AVL trees in order to free up memory. Error handling, such as input validation, was added in several places.